requirements.txt
streamlit
yfinance
pandas
plotly
pyyaml
requests
urllib3
